Added 'charset' parameter to checkStringValue().
Added optional character set parameter to static function checkStringValue()
in order to make sure that only ASCII and Latin-1 strings are checked for
their conformance with the VR definition (this is a current limitation of the
underlying VR checker). Also the checkValue() method now uses the value of
the SpecificCharacterSet element of the surrounding dataset/item.