Added new variant of checkElementValue().
This new variant gets a pointer to a DICOM element (and not a reference).
In addition, the getAndCheckXXXFromDataset() methods now make sure that no
element copy is created _before_ the element is checked. This is because
otherwise there would be no access to the SpecificCharacterSet element of
the surrounding dataset/item, which is needed for the VR checker.