Fixed wrong/missing comment on method parameter.
Fixed wrong comment on parameter "writeXfer" for the saveFile() method. In
case of EXS_Unknown not the currently selected transfer syntax is used but
the original one, i.e. the one in which the dataset was originally read or
created. Also added missing comment on parameter "oxfer" to write() methods.