Added OFFilename versions to some helper functions.
Added OFFilename versions to the helper functions getDirNameFromPath(),
getFilenameFromPath(), normalizeDirName(), combineDirAndFilename(),
searchDirectoryRecursively() and getFileSize() in order to enhance the
support for the Windows-specific wide character encoding (UTF-16).

Internally, the same code is used for both the OFString and OFFilename
version of the helper functions, which should facilitate finding any bugs.