Fixed issue with automatically generated classes.
Fixed issue with the detection of attributes nested in a sequence class (in
the automatically generated source code). Sometimes, there were additional
attributes associated with a sequence element, which actually belong to the
surrounding nesting level or to some other sequence element in the module.
Also changed VM for Irradiation Event UID (0008,3010) from "1" to "1-n".