Changed return type of addContentItem(node*).
Changed return type of addContentItem() method that expects a pointer to a
newly created document tree node. It now returns an OFCondition (like other
similar methods) instead of the pointer to the tree node that was passed to
it. Also fixed possible memory leaks in case adding the new tree node failed.