Further enhanced by-reference relationships.
Now storing the value type of a referenced node (target content item) in the
by-reference tree node class in order to facilitate IOD constraint checking.
Added new methods to by-reference tree node class in order to avoid direct
access to the member variables (by a friend class), and to allow updating the
referenced node. Added missing support for by-reference relationships in
DSRDocumentSubTree::checkSubTreeConstraints(). Also added a new test on this.