Updated "dcmrt" classes based on DICOM 2014b.
Updated automatically generated IOD and sequence c++ classes for the various
RT objects based on the current edition of the DICOM standard (2014b). This
time, the official DocBook/XML version has been be used (previously, it was
an unofficial version of the DICOM 2007 standard edition).
Please note the "known issues" that are documented in the module description.