Several fixes regarding OFnumeric_limits.
The initialization state of FLT_ROUNDS is now dumped to DCMTK_ROUND_STYLE
during configuration, which can then be evaluated in a compile time
context. This is necessary to conform closely enough to the C++ standard
as required by clang.
Mac OS X's special floating point exception control macros are now applied
on that platform, instead of the GNU POSIX extensions previously employed
there as well.
The existence of '_fpclassf' on the Windows platform is now probed
regarding the linker as well (previously only the compile time presence was
tested).