Introduced cross compiling support for Android.
Worked around some limitations when building for Android, for instance by
creating some configure tests that detect the absence of some operating
system functionality and enable existing or new fallback implementations in
their stead.
Added assisted cross compiling support to the CMake scripts, using a
specially adjusted version of an Android toolchain configuration file
borrowed from the OpenCV project. Also adding some code to (remote-)
control the Android device emulator (that is part of the official Android
SDK) and integrate it into the CMake build system. This allows the
configure- and unit tests to run on the emulated target device.