Optionally accept invalid content item values.
If a certain read flag is set, further violations of the DICOM standard are
accepted when reading an invalid SR document. For example, when the numeric
value of a NUM content item violates the maximum length. This might be useful
in order to make sure that the other information of the content item is read
(like the measurement unit).