Added support for long code values (CP-1031).
Added initial support for long code values according to CP-1031. By default,
the type of a code value (short, long or URN) is determined automatically,
e.g. when setting the value of a DSRCodedEntryValue instance. However, the
type can also be specified manually. Please note that for some methods, the
position of the optional "check" parameter moved after the new "codeValueType"
parameter. Support for the optional Equivalent Code Sequence (0008,0121) is
still missing.