Fixed issue with addNode() in basic tree class.
Fixed issue with addNode() method in basic tree class: When adding a new node
before or after the current node, and this new node has siblings (i.e. it is
actually a subtree without a dedicated root node), the connection to the
existing tree was not always correct. Also added test cases for this purpose.