Added base class for context groups.
Added base class for context groups. This is for the initial support only.
Specific context groups from the DCMR (DICOM Content Mapping Resource) will
follow. The class supports an arbitrary list of coded entries (if the group
is extensible); it also supports the selection of a particular entry.

Acknowledgement: This work and also the previous commits that are related
to the "dcmsr" module have been supported in part by the "QIICR" project.