Allow for disabling check on "DICOM file format".
Added new mode that allows for disabling the check on the "DICOM file
format", i.e. whether the input file contains file meta information
(as it is required by part 10 of the standard for all DICOM files).

As it is true for all disableXXXCheck() methods, please use this new mode
with care. It has been introduced to allow for registering files without
meta-header to a DICOMDIR, not to be used on a DICOM storage medium but
for local storage (e.g. as an index for all "DICOM files" that reside in
a certain directory of the filesystem).