Disable built-in dictionary for CMake Unix builds.
On Unix, DCMTK is traditionally built without a built-in dictionary that is
compiled into the binaries, using automake. Instead, the dictionary is read
during runtime from a well-known file under the installation directory.

However, this behaviour was not correctly adapted when porting the existing
CMake build toolchain from Windows to Unix so that using CMake under Unix also
resulted in a built-in dictionary. This commit makes sure CMake under Unix now
shows the same behaviour as the existing automake build. In the near future we
will also add a CMake option to toggle between built-in and file-based
dictionary.