Refactored handling of some <cmath> functions.
Added configure tests to detect wether isinf(), isnan() and finite() are
declared inside the std namespace or in the global scope. Updated the
implementation of OFStandard::isinf() and OFStandard::isnan() respectively.
Moved implementation to new file ofmath.cc and shared this implementation
for generating arith.h.
This fixes problems on OpenBSD and FreeBSD.