Enhanced handling of empty Measured Value Sequence.
Enhanced handling of Numeric Measurement Values being absent (i.e. both
Numeric Value and Measurement Unit have an empty value).  This is valid
according to the DICOM standard and should, therefore, also be supported
when calling the isValid() or setValue() methods of the class managing
Numeric Measurement Values in an SR document.  It is recommend but not
required to specifiy the reason for the absence of the Measured Value
Sequence (where these values are stored in a DICOM dataset) in the optional
Numeric Value Qualifier Code Sequence (which is also supported, of course).

Also added new test that makes sure that the internal processing is correct.

Acknowledgement: This work has been supported in part by the "QIICR" project.