Introduced new modules dcmiod, dcmfg and dcmseg.
dcmiod is a module that makes it easy to work with common DICOM IOD modules
and IODs being composed of them. When writing modules it is ensured that
basic checks are performed regarding the presence of required attributes and
their values as well as the validity of values. Classes like DcmIODCommon and
DcmIODImage can be used to derive user-defined classes that implement specific
DICOM IODs, making it easier to concentrate on the specifics of a particular
IOD.

The dcmfg module facilitates working with functional group that have been
introduced to DICOM in the context of the "Enhanced" family of IOD objects
that contain multiple frames that are described via shared and per-frame
functional groups. This class offers a generic interface to create, modify,
load and store such functional groups. Also, it supports a dozen of
functional groups directly by offering classes with related getter and setter
methods.

Finally, the dcmseg module makes use of dcmiod and dcmfg in order to
implement an Enhanced DICOM object, the Segmenation IOD. Since most of the
underlying data elements are already provided by dcmiod and dcmfg, dcmseg
only consists of a single class and some helpers.

Acknowledgement: This work has been supported in part by the "QIICR" project.
QIICR is supported by NIH National Cancer Institute, award U24 CA180918.