Made sure that IOD constraint checker is used.
Made sure that constraint checker for the Radiopharmaceutical Radiation
Dose SR IOD, which was introduced with commit 365951ec, is actually used.
In order to make sure that a constraint checker will never be forgotten
again for a newly supported SR document type, the "default" case has been
replaced in the corresponding switch() statement.