Build flags for internal and external dictionary.
Introduced build flags for controlling whether DCMTK will use an internal,
compiled-in and/or an external dictionary. On Windows systems, a
built-in dictionary (only) is the default, while on Unix only the external one
is enabled per default. This behaviour is identical to the old one. It is
possible to enable both options simultanously; then the built-in is loaded
first, and the external one is loaded afterwards. More details can be found
in dcmdata/docs/datadict.txt

Acknowledgement: This work has been supported by the "QIICR" project.