Added initial version of SRT and UCUM codes.
Added initial version of SNOMED-RT (Coding Scheme "SRT") and Unified Code for
Units of Measure ("UCUM") code definitions. In contrast to the "DCM" codes,
which were added only recently to the DCMTK, these two new coding schemes only
consist of those codes that are needed for TID 1600 to 1607 (Image Library).
Further codes will be added when needed.

The code definitions are generated semi-automatically from the machine-readable
DocBook/XML version of DICOM PS 3.16 (DICOM Content Mapping Resource). That
means that the definitions will be updated from time to time as the standard
evolves. However, please be warned that the names of the constants/defines
might change in the future (hopefully, there will be official keywords in the
standard like they are for the DICOM attribute tags in PS 3.6).

Acknowledgement: This work has been supported in part by the "QIICR" project.