Added initial version of TID 1600 (Image Library).
Added initial version of TID 1600 (Image Library) and included templates
TID 1601 to 1607. Also added a quite comprehensive test for this new template.
Please note that there are still a few details in the source code that are
marked with "tbc" (to be checked) or "tbd" (to be done). Also the class
documentation contains corresponding notes and warns about this fact.

Acknowledgement: This work has been supported in part by the "QIICR" project.