Added support for annotating content items.
Added support for textual annotations, which can be used to label content
items with user-defined information. This might facilitate the navigation
within the document tree and is especially useful for document structures
based on templates. Therefore, the template classes in the "cmr" (Content
Mapping Resource) submodule annotate all content items with a predefined
text of the format "TID xxxx - Row nn".