Made sure that tree node annotations are copied.
Made sure that tree node annotations are copied/cloned when a copy or clone
of a tree node or a (sub)tree is created. This makes sure that the additional
textual information (e.g. a user-defined label or the template ID/row number)
is also available in the copy/clone. As always: enhanced existing test cases.