Check return value of listen() for errors:
Introduced check of return value of listen() call in order to detect e.g.
when listening on a certain port that is already in use (EADDRINUSE).
This kind of error has been reported for various Linux distributions,
e.g. RHEL5.3 and CentOS 7.

Thanks to Thomas Sondergaard <thomas.sondergaard@karoshealth.com> and
Alexander Karaivanov <alexander.karaivanov@karoshealth.com> for the report
and suggested fix.