Handle EINTR raised by connect() or select().
The EINTR error (system call was interrupted by a signal that was caught)
might also be raised by a call of connect() or select(), which was not yet
handled appropriately.

Thanks to Thomas Sondergaard <thomas.sondergaard@karoshealth.com> and
Alexander Karaivanov <alexander.karaivanov@karoshealth.com> for the report
and suggested fix.