Added support for replacing a node in a tree.
Added support for replacing the current node (and all of its child nodes)
of a tree by a given node (which might be the root of a subtree). This new
feature is needed for the upcoming enhanced SR template support.

Acknowledgement: This work has been supported in part by the "QIICR" project.