Added 1st version of TID 1500 (Measurement Report).
Added first version of TID 1500 (Measurement Report) and included templates.
The focus of this initial commit was on the general structure of this SR
template and the mandatory content items and included templates (TID 1204,
1001, 1600). There is also a new test case that demonstrates how to use the
new TID1500_MeasurementReport class. Further enhancements will follow.

Acknowledgement: This work and also the previous commits on the "dcmsr"
module have been supported in part by the "QIICR" project. See CREDITS file
for details.