Added preliminary support for TID 1500 - Row 12.
Added preliminary support for the conditional CONTAINER content item
"Qualitative Evaluations" (TID 1500 - Row 12), which seems to be required
according to the condition given in the template definition table. Support
for the child content items (CODE and TEXT) will follow in a separate commit.

Acknowledgement: This work has been supported in part by the "QIICR" project.