Fixed bug when packaging binary segmentations.
Segmentations bits are filled into the byte from right to left, i.e. starting
from the least significant bit of the first byte to the most significant bit
of the last byte in the Pixel Data element. The old implementation filled the
bytes in the correct order, however, the bits have been filled in from the
wrong order. This was discovered since another implementation was available for
interoperability testing.

Acknowledgement: This work has been supported by the "QIICR" project.