Fixed another issue with createExpandedSubTree().
When an empty "included template" content item was processed, i.e. removed
from the resulting document tree, the next content item was skipped by
mistake. This was detected when enhancing a test case for the upcoming
support of TID 1411 (which is included in TID 1500 - Row 8).

Acknowledgement: This work has been supported in part by the "QIICR" project.