Added initial version of NCIt code definitions.
Added initial version of the NCI Thesaurus (Coding Scheme "NCIt") code
definitions as used/defined by the DICOM standard.

The code definitions are generated semi-automatically from the machine-readable
DocBook/XML version of DICOM PS 3.16 (DICOM Content Mapping Resource). That
means that the definitions will be updated from time to time as the standard
evolves. However, please be warned that the names of the constants/defines
might change in the future (hopefully, there will be official keywords in the
standard like they are for the DICOM attribute tags in PS 3.6).

Acknowledgement: This work has been supported in part by the "QIICR" project.