Enhanced check on Segmentation objects.
Enhanced check on DICOM Segmentation objects that are referenced by an
instance of DSRImageReferenceValue. On the one hand, the list of known
Storage SOP Classes has been updated for the latest edition of the
DICOM standard (2015c); the Surface Segmentation object was missing.
On the other hand, the list of referenced segment numbers is only allowed
for Segmentation objects, which is now also checked.

Acknowledgement: This work has been supported in part by the "QIICR" project.