Defined short names for CID and TID classes.
Defined short names for context group (CID) and template (TID) classes
as an alternative to the existing long names. The nomenclature of these
type definitions is: CMR_CIDnnn (with "nnn" being the numeric identifier
of the context group) or CMR_TIDnnn (with "nnn" being the numeric ID of
the SR template), respectively.

Acknowledgement: This work has been supported in part by the "QIICR" project.