Better logging of select() on network.
When waiting for network data using select() it has not been distinguishable
whether a regular timeout occurred or if another error (such as an interrupt)
interrupted the call. Now, debug output is written in order to make it easier
to see what exactly happened (especially in case of errors).

Thanks to forum user michael12345 for the report.