Changed default value of "expandTree" parameter.
The default value of the "expandTree" parameter of the
DSRDocument::setTreeFromRootTemplate() method changed from OFFalse to OFTrue.
This is because some output methods like write() or renderHTML() do not yet
work on document trees that contain "included template" nodes.

Acknowledgement: This work has been supported in part by the "QIICR" project.