Added support for TID 1411 and included templates.
Added initial support for TID 1411 (Volumetric ROI Measurements) and
included templates TID 1502 and TID 1419. This is the first generic SR
template that has been implemented for this toolkit, i.e. it has SR
template parameters that are directly mapped to C++ template types.
Current support for TID 1411 is focussed on the use within TID 1500
(Measurement Report), which will be added by the next commit.

Acknowledgement: This work has been supported in part by the "QIICR" project.