Added new variant of setReferencedSegment().
Added new variant of adding/setting a referenced segment to the template
(TID 1411 - Row 7). According to CP-1494 the Tracking ID and Tracking UID
should be consistent with the corresponding element values of the referenced
DICOM dataset (if present).

Acknowledgement: This work has been supported in part by the "QIICR" project.