Fixed copying of tracking information (TID 1411).
Fixed copying of tracking information from a referenced segmentation object:
According to CP-1496 the Tracking ID and Tracking UID are stored in an item
in the Segment Sequence, so further code is needed to find the appropriate
item. Also adapted the test case accordingly.

Acknowledgement: This work has been supported in part by the "QIICR" project.