Fixed issue with addPersonObserver() in TID 1001.
Fixed issue when calling addPersonObserver() in TID1001_ObservationContext
for the first time _after_ addDeviceObserver() has already been called.
According to the SR template specification, the person observers always have
to be stored at the beginning of the subtree of the SR template, i.e. the
root node changes in such a case. Modified the test case accordingly.

Also see previous commit that fixes an issue on the basic tree class.

Acknowledgement: This work has been supported in part by the "QIICR" project.