Fixed issue with getMACAddress() on Windows 8++.
Fixed issue with getMACAddress() on some Windows 8 (and newer) systems.
Removed unneeded call to SnmpExtensionQuery() that caused a crash on such
systems.

Thanks to forum user "andreasb" for the report and suggested fix, and to
forum user "Rich in Soquel" for confirming that this fix also solves an
issue on Windows 10.

Closes DCMTK Bug #594 (hopefully)