Use addContentItem() for by-value relations only.
Made clear that addContentItem() should be used for by-value relationships
only. This is now also checked and a new test case makes sure that it works.
Furthermore, do not accept the internal value type VT_includedTemplate for
this method. All this is checked by the helper method canAddContentItem().

Acknowledgement: This work has been supported in part by the "QIICR" project.