Changed print() output of included SR templates.
Slightly changed the output of the print() method for "included template"
tree nodes: The line indentation and output of the position counter is now
also shown for this "internal" node (and not only for the included subtree);
the position counter is increased by 1 as it is true for all other tree
nodes, which makes it easier and more consistent when determining the target
of by-reference relationships. However, there is still some work to do...

Also, the separate line for the included SR template (starting with the text
"# INCLUDE ...") is now printed by default. Therefore, the name of the print
flag changed from PF_printXXX to PF_hideXXX.

Acknowledgement: This work has been supported in part by the "QIICR" project.