By default, convert wide char encoding to UTF-8.
Changed default value of parameter "convert" for both the constructor and
the set() method of the OFFilename class, which expects a "wchar_t *", from
"OFFalse" to "OFTrue". That means, a character string that is passed with
Windows UTF-16 encoding will be converted to UTF-8 (char *) encoding. This
change of the default behavior makes sense since e.g. the output of an
OFFilename instance to the logger always happens with UTF-8 encoding.