Overwrite isShort() in DSRIncludedTemplateTreeNode.
Overwrite isShort() method in class DSRIncludedTemplateTreeNode. Since the
base class DSRDocumentTreeNode always returns OFTrue, it makes sense to
check the content in the derived tree node class. Of course, rendering a
document tree with "included templates" is currently not supported, so the
isShort() method is probably never really used in this context.

Acknowledgement: This work has been supported in part by the "QIICR" project.