Fixed use of checkStringValue().
For some VR classes (DcmLongText, DcmShortText, DcmUnlimitedText) the method
checkStringValue() expects the character set as the second parameter and not
the VM as it is the case for most other VR classes. The related calls in
dcmiod, dcmfg and dcmseg did not respect this deviating API, thus requiring
the given fix which disables the value checking.