Enhanced check for invalid characters.
Now, the static function checkStringValue() - and all methods that use it -
check for invalid characters in the given value. Currently, this only works
for ASCII strings (default character repertoire) that contain non-ASCII
characters, i.e. character codes > 127.

This change requires that the Specific Character Set value is passed to the
respective function/method since the default is ASCII (as it was before).