Added methods to get external cursor to root node.
Added two methods that allow for getting an (external) cursor to the root
node of the document tree, either an instance of DSRDocumentTreeNodeCursor
or an instance of the recently introduced DSRIncludedTemplateNodeCursor.
These cursors can be used to iterate over the nodes of the document tree
without changing the internal cursor.

Acknowledgement: This work has been supported in part by the "QIICR" project.