Fixed issue with shared library build (MSVC++).
Fixed issue with shared library build on Windows with Microsoft Visual C++.
Now, the CMake option BUILD_SHARED_LIBS works again on both Linux with gcc
and Windows with Visual C++, i.e. for ".so" and ".dll" files.

C++ templates with explicit specialization and shared library support are
really a nightmare, especially when the implementation of the C++ template
class is not completely contained in a header file.