Updated data dictionary for DICOM 2016a edition.
Updated data dictionary for the latest edition of the DICOM standard, which
has been released on 2016-02-19. This includes the new tags from Supplement
156, 181 and 184 as well as those from CP 1364, 1431, 1432, 1441, 1454,
1457, 1458, 1460, 1461, 1478 (WC), 1487, 1496, 1499, 1504 and 1516. Please
note that some attributes changed their names and/or value multiplicity.

In addition, new and modified tags from ASTM's DICONDE (Digital Imaging and
Communication in Nondestructive Evaluation) standard have been incorporated.

The total number of entries in the DICOM data dictionary is now 4032.