Added comment on mapping of code meanings.
Added comment on the mapping of code meanings to a compiler-friendly
representention, which is the reason why the SRT and UCUM coding schemes
are not yet complete (with regard to the definitions in DICOM PS 3.16).

Please note that the commit message for the previous commit (168662c)
was incorrect: not the UCUM but the UMLS codes definitions were updated.

Acknowledgement: This work has been supported in part by the "QIICR" project.