Fixed possible issue in OFStandard helper function.
Fixed possible issue in OFStandard::removeRootDirFromPathname() that could
occur when passing a NULL pointer as the "rootDir" parameter, and thus
calling strncmp() or wcsncmp() with an invalid value. This is mainly a
theoretical scenario, since most implementations of these standard string
functions will not crash when a 0 length value is passed at the same time
(see test case "ofstd_OFStandard_removeRootDirFromPathname").

This possible issue has been found with Pareon Verify.

Thanks to Andrei Terechko <andrei@vectorfabrics.com> for the report.