Check for PixelSpacing in SharedFunctionalGroups.
Also check for the Pixel Spacing (0028,0030) attribute in the Pixel Measures
Sequence (0028,9110) of the Shared Functional Groups Sequence (5200,9229).
By doing so, non-square pixels are supported for the family of Enhanced
Multiframe Image IODs, too.

Please note that support for Pixel Spacing within the Per-frame Functional
Groups Sequence (5200,9230) is still missing.