Removed definition of Visual Studio macro NOMINMAX.
A previous commit introduced the definition of NOMINMAX inside osconfig.h to
fix problems regarding the min()/max() functions declared as part of
OFnumeric_limits. However, this is problematic for users that declare
NOMINMAX themselves or don't want to declare NOMINMAX at all.
This commit removes the NOMINMAX definition from osconfig.h and modifies
OFnumeric_limits and related header files in a way that they may be included
without the need to declare NOMINMAX. The NOMINMAX macro is then only defined
when compiling the DCMTK itself, DCMTK users may or may not declare NOMINMAX
themselves as needed/appropriate.
Thanks to Peter Klotz <Peter.Klotz@ith-icoserve.com> for the report.