Enhanced return value of determineStartFragment().
Enhanced error conditions returned by DcmCodec::determineStartFragment().
This avoids getting a meaningless EC_IllegalCall when trying to access a
particular frame of a compressed and fragmented pixel data element, e.g.
if the basic offset table is empty. This is particularly reasonable since
the codec rountines are usually hidden to the user of the toolkit, e.g. by
the DicomImage class.