Added size of DICOM dataset to transfer entry.
Added information on the size of the dataset (in bytes) to the transfer entry,
i.e. to the list of SOP instances to be transferred. Also added a new notifier
method that is called _before_ a SOP instance is transferred. This allows for
implementing a more accurate progress bar in a derived class.