movescu now returns non-zero if c-move was unsuccessful.
The movescu command line application now returns non-zero if the
status of one or more c-move responses is warning or error and also
prints a warning message to the logger. This simplifies the use of the
tool in shell scripts, which can check for the return code.

This closes feature request #652.