Added new option --socket-timeout to storescu/scp.
Added new option --socket-timeout, which allows for specifying the timeout
for sending and receiving data on a network socket. A default of 60 seconds
is useful in cases where the sender (e.g. storescu) looses the connection
to the receiver (e.g. storescp) because the network cable is pulled or the
like. A value 0 means that the send() or recv() function never timeouts.