Changed type for index/number of sequence items.
Changed data type used for the index and number of sequence items. The type
"size_t" seems to be more appropriate than "unsigned long" since internally
an OFList (subset of std:list) instance is used for the data management.

As a consequence, VisualStudio and other compilers will not warn anymore on
a "possible loss of data" (warning C4267).