Made sure that a value is loaded before modified.
Made sure that an element value is actually loaded before it is modified
with DcmElement::changeValue(). This fixes a possible crash (segmentation
fault) in case the "maxReadLength" parameter of the various loadFile/read
methods was set to a very low value (i.e., less than 4).

Thanks to Michael Schinner <Michael.Schinner@ith-icoserve.com> for the bug
report and for the test program that demonstrates this crash.