Report error in case of wrong Pixel Data encoding.
It has been reported that some DICOM datasets do write encapsulated pixel data
without following the related rules, i.e. to use undefined length for the Pixel
Data element and to use Pixel Sequence and Pixel Items inside.

So far such files (using defined length on encapsulated Pixel Data) have been
read without any warnings or errors giving the impression that the data is
valid.

Now, DCMTK will per default report an error if reading such files. There is a
global switch which permits to enforce the old behaviour, i.e. to read such
files as if the encoding would be correct. However, in that case a warning
is printed.

Options demonstrating the flag have been added to dcmdump and dcmconv.

The new functionality is aware of the fact that Pixel Data not being part of
the top level dataset, i.e. nested within sequences, must not use the
encapsulated format but remain uncompressed (as for icon images).

The validity of internal structures of the Pixel Data element (Pixel Sequence
and Items) is not checked by this new compatibility flag.

This closes DCMTK Bug #427.