Further enhanced support for non-patient objects.
Added optional parameter to dcmIsaStorageSOPClassUID() that allows for
checking a given SOP Class UID for being a known Storage SOP Class UID
of a non-patient DICOM object. Also added support to check for image
objects only. The default behavior of this function has not changed,
i.e. only those Storage SOP Classes are covered that fit into into the
conventional PATIENT-STUDY-SERIES-INSTANCE information model.

The function dcmIsImageStorageSOPClassUID() is now only a shortcut for
calling dcmIsaStorageSOPClassUID() with the new type parameter being
ESSC_Image. This is useful in order to be backward compatible.