Allow for sending non-patient DICOM objects.
The storescu and dcmsend tool (based on the class DcmStorageSCU) are now
able to send non-patient DICOM objects. The latter was already able to do
so but reported a warning to the DEBUG logger that an "unknown storage SOP
class (with standard UID prefix)" is to be transferred. This warning has
now disappeared.

This partly closes DCMTK Conformance #695.