Added two missing SOP Classes to modality table.
Added two missing non-patient Storage SOP Classes to the modality table,
which e.g. maps a known UID to a short character code ("modality") that
can be used to generate a meaningful filename prefix. That way, a storage
receiver of a Color Palette Storage object will generate a filename
starting with "CP." instead of "UNKNOWN." (which is the default in case
of an unknown UID).