Hide public element constructors setting length.
DCMTK's VR classes allowed to construct new DICOM elements by tag and a
given length value. Internally the length is set, however, no related memory
is allocated. This is the desired behaviour when reading new DICOM datasets
since after creation of the element with a specific length after reading tag
and length from the stream, the value is read and set on the element right
afterwards.

However, if such an element is created manually by a user who tries afterwards
to access (read or write) a value to that element, a crash can occur since
the element itself expects to have memory allocated up to the given length.

In order to fix this, the related VR classes' constructors have been removed
from the public interface and have been made protected. So far, only DcmItem
and partly DcmDirectoryRecord (as "parsing" classes constructing elements in
the valid way described above) need direct access to those constructors and
thus have been amended by related friend declarations.

The global functions newDicomElement(...) that use those constructors directly
or indirectly have been moved into DcmItem as static methods. The ones that
can be used publicly (those only taking tag and not length), are made public
while the only one taking length (and other parsing parameters) is protected
and is only used by DcmItem or safely from within other public methods.

An alternative approach would have been to change the parsing behaviour
that constructs element with explicit length first, in order to fill them
with a value in the second step. Instead, onld could construct with 0
length and only set the length when actually setting the value. This
would make the explicit length constructors superfluous.
However, since the parser's internals are quite complicated regarding
length handling, it was not touched for now. However, the work done may make
it easier to change the parser accordingly later.

Another approach would be to always allocate memory in the constructors.
However this will lead often to double initializations since most methods
that set a value in a subsequent call will delete that memory to allocate
the value newly a second time. Also, the possibility to create elements with
an explicit empty value of a specific length is of limited use: For number-
based VRs this can easily be imitated by calling the related set() method
afterwards, while for strings it is not even clear what an empty value
of a specifc length might be.

Thanks to forum user spasmous <spasmous@gmail.com> for the report.

This fixes DCMTK bug #374.