Fixed crash while processing an invalid DICOMDIR.
Fixed a crash that could occur while processing an invalid DICOMDIR file,
e.g. when a single Directory Record was referenced more than once and,
therefore, deleted more than once (from the DcmDicomDir destructor). Now,
an error is both reported to the logger and returned to the caller.

The main changes were done to DcmDicomDir::moveRecordToTree(). Additionally,
slightly enhanced comments on other functions and improved the logger output.

Thanks to forum user "psih128" for the report and for the sample file.

Closes DCMTK Bug #388.