Refactored string to date and time conversion.
Refactored DcmDate, DcmTime and DcmDateTime string parser functions to work
on character string + size input in addition to the existing functions that
work on OFStrings. The new functions do not require the strings to be zero
terminated and are therefore composable.
DcmDateTime::getOFDateTimeFromString is now implemented based on
DcmTime::getOFTimeFromString, which is now possible since the functions
can now be used on substrings without unnecessary copies.
Default parameters have been replaced by respective overloads to provide
a consistent API in all three classes (all now have a function that takes
no additional parameters to the string and the result).