Added character set conversion support to dcmqrdb.
The dcmqrdb library now supports on the fly character set conversion for
attribute matching and converting the find response to the requested character
set.
The matching has been refactored to use DcmAttributeMatching internally.
Increased the size of attributes that are affected by Specific Character Set
in the index file so that values using multi byte encodings may still fit.